refactor: merge single-consumer modules
