--- src/fsync.ts
@@ -1,1 +1,0 @@
-export { syncDirectoryBestEffort } from "./directory-durability.js";
