--- src/mode.ts
@@ -1,3 +1,0 @@
-export function formatPosixMode(mode: number): string {
-  return (mode & 0o777).toString(8).padStart(3, "0");
-}
