+++ src/root-errors.ts
@@ -74,3 +74,13 @@ export function normalizeRemovePathError(error: unknown): Error {
   }
   return new FsSafeError("not-removable", "path could not be removed", { cause });
 }
+
+export function throwFsSafeReadError(error: unknown, label: string): never {
+  if (error instanceof FsSafeError) {
+    throw error;
+  }
+  if (isNodeError(error)) {
+    throw new FsSafeError("read-failed", `${label} target could not be read`, { cause: error });
+  }
+  throw error;
+}
